Update apigateway.md
configuration syntax mistake in request schema validator , schemas is changed to schema in latest update of serverless-offline. previously i was getting error --> errorMessage": "Expected `schema` to be an object or boolean", "errorType": "SchemaError", "offlineInfo": "If you believe this is an issue with serverless-offline please submit it, thanks. https://github.com/dherault/serverless-offline/issues", "stackTrace": [ "SchemaError: Expected `schema` to be an object or boolean", "at Validator.validate.